socket/requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
dotenv==0.0.5
idna==3.6
rel==0.4.9.1
requests==2.31.0
urllib3==2.1.0
websocket-client==1.7.0
